# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CreateKegiatanRequest:
    properties:
      deskripsi:
        type: string
      judul:
        type: string
      kategori_id:
        type: integer
      tanggal_berjalan:
        type: string
      tanggal_diajukan:
        type: string
    type: object
...
info:
  contact: {}
  description: Dokumentasi API backend Sistem Informasi Karang Taruna
  title: API Sistem Informasi Karang Taruna
  version: "1.0"
paths:
  /api/kegiatan:
    post:
      consumes:
      - application/json
      description: Ketua divisi membuat pengajuan kegiatan
      parameters:
      - description: Create kegiatan payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateKegiatanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create kegiatan
      tags:
      - Ketua Divisi Kegiatan
  /api/kegiatan-saya:
    get:
      description: Mengambil daftar kegiatan milik ketua divisi
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search judul
        in: query
        name: search
        type: string
      - description: Kategori ID
        in: query
        name: kategori
        type: string
      - description: 'Status: pending, approved, rejected'
        in: query
        name: status
        type: string
      - description: 'Sort: terbaru atau terlama'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get kegiatan saya
      tags:
      - Ketua Divisi Kegiatan
  /api/kegiatan/{id}:
    delete:
      description: Menghapus kegiatan milik ketua divisi selama status masih pending
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete kegiatan
      tags:
      - Ketua Divisi Kegiatan
    put:
      consumes:
      - application/json
      description: Mengubah data kegiatan milik ketua divisi selama status masih pending
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      - description: Data update kegiatan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateKegiatanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update kegiatan
      tags:
      - Ketua Divisi Kegiatan
...
  /api/ketua/kegiatan:
    get:
      description: Mengambil semua kegiatan untuk ketua umum
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search judul
        in: query
        name: search
        type: string
      - description: Kategori ID
        in: query
        name: kategori
        type: string
      - description: 'Status: pending, approved, rejected'
        in: query
        name: status
        type: string
      - description: 'Sort: terbaru atau terlama'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all kegiatan for ketua umum
      tags:
      - Ketua Umum Kegiatan
...
  /api/ketua/kegiatan/user/{id}:
    get:
      description: Mengambil kegiatan berdasarkan user tertentu
      parameters:
      - description: ID User
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get kegiatan by user
      tags:
      - Ketua Umum Kegiatan
  /api/notifications:
    get:
      description: Mengambil semua notifikasi milik user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get notifications
      tags:
      - Notifications
  /api/notifications/{id}/read:
    patch:
      description: Menandai notifikasi sebagai sudah dibaca
      parameters:
      - description: ID Notification
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Read notification
      tags:
      - Notifications
...
  /kegiatan:
    get:
      description: Mengambil daftar kegiatan publik yang berstatus approved
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Search judul
        in: query
        name: search
        type: string
      - description: Kategori ID
        in: query
        name: kategori
        type: string
      - description: 'Sort: terbaru atau terlama'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all public kegiatan
      tags:
      - Public Kegiatan
  /kegiatan/{id}:
    get:
      description: Mengambil detail kegiatan publik berdasarkan ID
      parameters:
      - description: ID Kegiatan
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get detail kegiatan
      tags:
      - Public Kegiatan
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"